# docs/swagger.yaml
basePath: /api
definitions:
...
  database.Tenant:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      external_ref:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/database.TenantStatus'
      updated_at:
        $ref: '#/definitions/pgtype.Timestamptz'
    type: object
...
  database.TenantStatus:
    enum:
    - active
    - suspended
    - deleted
    type: string
    x-enum-varnames:
    - TenantStatusActive
    - TenantStatusSuspended
    - TenantStatusDeleted
  database.UsageMetricsDaily:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      endpoints_active:
        type: integer
      eps_avg:
        type: integer
      eps_peak:
        type: integer
      events_ingested:
        type: integer
      id:
        type: string
      tenant_id:
        type: string
      usage_date:
        $ref: '#/definitions/pgtype.Date'
    type: object
...
  handlers.CreateTenantRequest:
    properties:
      external_ref:
        example: ext-acme-123
        type: string
      name:
        example: Acme Corp
        type: string
      status:
        example: active
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.UpdateTenantStatusRequest:
    properties:
      status:
        example: suspended
        type: string
    type: object
  handlers.UpsertUsageRequest:
    properties:
      endpoints_active:
        example: 150
        type: integer
      eps_avg:
        example: 1736
        type: integer
      eps_peak:
        example: 4200
        type: integer
      events_ingested:
        example: 1500000
        type: integer
      usage_date:
        example: "2026-03-07"
        type: string
    type: object
...
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: Cloud License Platform — multi-tenant SaaS licensing and entitlement
    management.
  title: CLP API
  version: "1.0"
paths:
...
  /tenants:
    get:
      description: Returns a paginated list of tenants ordered by creation time descending.
      parameters:
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Tenant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Creates a new tenant. Status defaults to "active" if omitted.
      parameters:
      - description: Tenant
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Tenant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      description: Permanently deletes a tenant by UUID.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a tenant
      tags:
      - tenants
    get:
      description: Returns a single tenant by UUID.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Tenant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a tenant
      tags:
      - tenants
...
  /tenants/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets the tenant status (active, suspended, deleted).
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTenantStatusRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update tenant status
      tags:
      - tenants
...
  /tenants/{id}/usage:
    get:
      description: Returns daily usage metrics for the given tenant between from and
        to dates (inclusive), newest first.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.UsageMetricsDaily'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get usage metrics for a date range
      tags:
      - usage
    put:
      consumes:
      - application/json
      description: Creates or updates usage metrics for the given tenant and date
        (idempotent).
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Usage metrics
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpsertUsageRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upsert daily usage metrics for a tenant
      tags:
      - usage
  /tenants/{id}/usage/{date}:
    get:
      description: Returns usage metrics for the given tenant on the specified date
        (YYYY-MM-DD).
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.UsageMetricsDaily'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get usage metrics for a specific date
      tags:
      - usage
  /tenants/external/{ref}:
    get:
      description: Looks up a tenant using its external system reference identifier.
      parameters:
      - description: External reference
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Tenant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get tenant by external reference
      tags:
      - tenants
swagger: "2.0"